feat: implement listener for real-time log updates

+++ app/src/main/java/kite1412/irrigo/model/LightIntensityLog.kt
@@ -4,7 +4,7 @@ import kotlin.time.Instant
 
 data class LightIntensityLog(
     val device: Device,
-    val lux: Int,
+    val lux: Double,
     val status: LightIntensityStatus,
     val timestamp: Instant
 )
